add delete method in usuario, alumno, tecnico and personal classes an…

+++ proyecto_satapp_Carlos_Rafa/src/main/java/com/example/proyecto_satapp_Carlos_Rafa/services/PersonalService.java
@@ -58,7 +58,12 @@ public class PersonalService {
 
     }
 
-    public void delete(Long id) {
+    public void deleteById(Long id) {
+        Optional<Personal> personalOp = personalRepository.findById(id);
+
+        if (personalOp.isEmpty()) {
+            throw new EntityNotFoundException("Personal no encontrado");
+        }
+
         personalRepository.deleteById(id);
-    }
-}
+    }}
